test: fixed race condition with endpoint deletion
client.endpoints().inNamespace(defaultNamespace).withName(serviceName).delete()

doesn't wait for the operation to complete in the cluster.

We either need to wait for a condition where these endpoint
is null.
Or use the convenience .withTimeout DSL method so that
Kubernetes client blocks the thread for us in a best effort
attempt to wait until the referenced Kubernetes object no longer
exists.